+++ config/routes.rb
@@ -12,7 +12,7 @@ RecordingStudioStripe::Engine.routes.draw do
 
   namespace :admin do
     resources :products, only: %i[new create edit update]
-    resources :prices, only: %i[new create]
+    resources :prices, only: %i[new create edit update]
     resources :meters, only: %i[new create]
     resources :paywalls, only: %i[new create]
   end